Rename: スタイルガイドに沿ってファイル名を変更
スネークケースで命名されていたコンポーネントファイル名をパスカルケースに変更
