+++ components/HideButton.vue
